migrate quickbar types
